requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
supabase==2.7.4
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
httpx==0.27.2
